+++ fabric/src/main/java/immersive_aircraft/fabric/ClientFabric.java
@@ -1,10 +1,12 @@
 package immersive_aircraft.fabric;
 
+import com.mojang.blaze3d.systems.RenderSystem;
 import immersive_aircraft.ClientMain;
 import immersive_aircraft.ItemColors;
 import immersive_aircraft.Renderer;
 import immersive_aircraft.WeaponRendererRegistry;
 import immersive_aircraft.client.KeyBindings;
+import immersive_aircraft.client.OverlayRenderer;
 import immersive_aircraft.item.upgrade.VehicleStat;
 import immersive_aircraft.item.upgrade.VehicleUpgrade;
 import immersive_aircraft.item.upgrade.VehicleUpgradeRegistry;
@@ -14,7 +16,10 @@ import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.item.v1.ItemTooltipCallback;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.rendering.v1.ColorProviderRegistry;
+import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import net.minecraft.ChatFormatting;
+import net.minecraft.client.DeltaTracker;
+import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -37,10 +42,24 @@ public final class ClientFabric implements ClientModInitializer {
         Renderer.bootstrap();
         WeaponRendererRegistry.bootstrap();
 
+        HudRenderCallback.EVENT.register(ClientFabric::renderOverlay);
+
         ItemColors.ITEM_COLOR_PROVIDERS.forEach((item, itemColor) -> ColorProviderRegistry.ITEM.register(itemColor, item.get()));
 
         KeyBindings.list.forEach(KeyBindingHelper::registerKeyBinding);
         ItemTooltipCallback.EVENT.register(this::itemTooltipCallback);
+    }
+
+    private static void renderOverlay(GuiGraphics guiGraphics, DeltaTracker deltaTracker) {
+        guiGraphics.flush();
+        guiGraphics.pose().pushPose();
+        try {
+            OverlayRenderer.renderOverlay(guiGraphics, deltaTracker.getGameTimeDeltaTicks(), 49);
+            guiGraphics.flush();
+        } finally {
+            guiGraphics.pose().popPose();
+            RenderSystem.setShaderColor(1.0f, 1.0f, 1.0f, 1.0f);
+        }
     }
 
     /**